Speeding up scanning.
Removed a bottleneck where a whole file was loaded when it didn't need
to be. Mind you, when scanning is doing more, then one might have to
load the whole file anyway. But that is for later.

Also added a "meta" route, to check what is in a page's meta-data.