Figured out why links were not being ignored.
In IkiWiki, the "preprocessing directive" part of the scanning phase
is run before the "linkify" part, and as part of the preprocessor
part, it removes the preprocessing directives from the content when
they've been processed. So they aren't there later when the linkify
is done. That makes sense, but it hadn't occurred to me.