Very crude "includepage" hook.
Rather than sophisticatedly processing all the included pages and then
adding them in later, this takes a naive approach, and simply
includes the raw content of the included pages VERY early in the
assembly pass. This only works if the pages are the same type...
but that's good enough for my needs right now.