Got rid of the globalinfo stuff.
I only put it there for my initial development of the shortcuts plugin,
and in the end it made more sense to me to put the defintions of the
shortcuts into the config file (like PmWiki) rather than into a page
(like IkiWiki). But I hadn't gotten around to removing it before now.